Add snaphot storer interface and kopia implementation
Snapstore package has two interfaces.
Storer describes the ability to store and retrieve a buffer
of metadata associated with a string key.
DataPersister describes the ability to flush a store of
metadata to a repository and load it again.

Add the kopia metadata repository implementation, which
implements the Storer and DataPersister interfaces.