Validation of snapshot list
- Adding an env-configurable limit on the number
of live snapshots that can be deleted from a repo
in a single engine run. Defaults to 10. Error will be
thrown instead of deleting more.
- Adding validation of the `manifest list` call by
comparing the length of the snapshot ID list
against the length of the list from `snap list
--all --manifest-id --show-identical`
- Adding unit test for parsing snap list all
- Adding kopia integration tests for snapshot ID
listing kopia commands to catch if the parsing
format ever changes.